Added types.h, variables.c/h
organized header organization a little bit

+++ functions.c
@@ -4,7 +4,12 @@
 #include <stdint.h>
 #include <string.h>
 
-#include "operator.h"
+#include "types.h"
+#include "matrix.h"
+#include "stack.h"
+#include "variables.h"
+
+#include "operatorUtility.h"
 #include "functions.h"
 
 //factorial function
@@ -26,8 +31,8 @@ matrix *eye(matrix *a, error_return *error){
 	uint16_t newSize[3] = {a->elements[0], a->elements[0], 0};
 	matrix *out = initMatrix(newSize, 2, error);
 
-	int index = 0;
-	for(int i = 0; i < a->elements[0]; ++i){
+	uint64_t index = 0;
+	for(uint64_t i = 0; i < a->elements[0]; ++i){
 		index = i * (a->elements[0] + 1);
 		out->elements[index] = 1;
 	}
@@ -45,7 +50,7 @@ matrix *getSize(matrix *a, error_return *error){
 	newSize[2] = 0;
   
 	matrix *out = initMatrix(newSize, 2, error);
-	for(int i = 0; a->size[i]; ++i){
+	for(uint64_t i = 0; a->size[i]; ++i){
 		out->elements[i] = a->size[i];
 	}
 
@@ -63,7 +68,7 @@ matrix *magnitude(matrix *a, error_return *error){
 		if((a->size[0] == 1) || (a->size[1] == 1)){
 			element magnitudeA = 0;
 
-			for(int i = 0; i < a->length; ++i){
+			for(uint64_t i = 0; i < a->length; ++i){
 				magnitudeA += (a->elements[i] * a->elements[i]);
 			}
 
@@ -97,7 +102,7 @@ matrix *numel(matrix *a, error_return *error){
 matrix *reference(matrix *a, matrix *b, error_return *error){
 	matrix *out = copyMatrix(b, error);
 
-	for(int i = 0; i < b->length; ++i){
+	for(uint64_t i = 0; i < b->length; ++i){
 		out->elements[i] = a->elements[(uint64_t) (b->elements[i])];
 	}
 
@@ -137,8 +142,8 @@ matrix *assign(matrix *a, matrix *b, vari *var, error_return *error){
 
 		} else{
 
-			for(int i = 0; i < var->assignIndex->length; ++i){
-				a->elements[(int) var->assignIndex->elements[i]] = b->elements[i];
+			for(uint64_t i = 0; i < var->assignIndex->length; ++i){
+				a->elements[(uint64_t) var->assignIndex->elements[i]] = b->elements[i];
 			}
 
 			incrementFlag = 0;
@@ -186,15 +191,15 @@ matrix *multiplyMatrix(matrix *a, matrix *b, error_return *error){
 	matrix *transposeA = transposeMatrix(a, error);
 
 	//counter for elements put into out
-	int l = 0;
+	uint64_t l = 0;
 
 	//generic O(n^3) algorithm
 	
 	//transpose a and then multiply every column with every other column in each matrix
-	for(int i = 0; i < b->size[1]; ++i){
+	for(uint16_t i = 0; i < b->size[1]; ++i){
 		for(int j = 0; j < transposeA->size[1]; ++j){
 			element tempSum = 0;
-			for(int k = 0; k < transposeA->size[0]; ++k){
+			for(uint16_t k = 0; k < transposeA->size[0]; ++k){
 				tempSum += transposeA->elements[k + j * transposeA->size[0]] * b->elements[k + i * b->size[0]];
 			}
 			out->elements[l] = tempSum;
@@ -208,8 +213,8 @@ matrix *multiplyMatrix(matrix *a, matrix *b, error_return *error){
 
 
 matrix *exponentMatrix(matrix *a, matrix *b, error_return *error){
-	int aScalar = isScalar(a);
-	int bScalar = isScalar(b);
+	uint8_t aScalar = isScalar(a);
+	uint8_t bScalar = isScalar(b);
 
 	matrix *tempMat = NULL;
 	matrix *out = NULL;
@@ -223,7 +228,7 @@ matrix *exponentMatrix(matrix *a, matrix *b, error_return *error){
 		if(aScalar){ //a is the scalar
 			out = copyMatrix(b, error);
 
-			for(int i = 0; i < out->length; ++i){
+			for(uint64_t i = 0; i < out->length; ++i){
 				out->elements[i] = pow(a->elements[0],b->elements[i]);
 			}
 
@@ -237,7 +242,7 @@ matrix *exponentMatrix(matrix *a, matrix *b, error_return *error){
 
 			if((b->elements[0] - floor(b->elements[0])) < 0.00000000001){
 				int64_t power = b->elements[0];
-				for(int i = 0; i < power; ++i){
+				for(uint64_t i = 0; i < power; ++i){
 					out = multiplyMatrix(tempMat, a, error);
 					freeMatrix(tempMat);
 
@@ -279,12 +284,12 @@ matrix *transposeMatrix(matrix *a, error_return *error){
 	matrix *out = initMatrix(newSize, 2, error);
 
 	//new index
-	int newInd = 0;
+	uint64_t newInd = 0;
 
 	//temp variable for the sublocation of the new index
-	int subLoc = 0;
+	uint64_t subLoc = 0;
   
-	for(int i = 0; i < out->length; ++i){
+	for(uint64_t i = 0; i < out->length; ++i){
 		//subLoc is an int and gets rounded down
 		subLoc = i/a->size[0];
 		newInd = subLoc + a->size[1] * (i - subLoc * a->size[0]);
@@ -300,7 +305,7 @@ matrix *transposeMatrix(matrix *a, error_return *error){
 matrix *min(matrix *m, error_return *error) {
 	element out = m->elements[0];
 
-	for(int i = 1; i < m->length; ++i) {
+	for(uint64_t i = 1; i < m->length; ++i) {
 		out = fmin(out, m->elements[i]);
 	}
 
@@ -312,7 +317,7 @@ matrix *min(matrix *m, error_return *error) {
 matrix *max(matrix *m, error_return *error) {
 	element out = m->elements[0];
 
-	for(int i = 1; i < m->length; ++i) {
+	for(uint64_t i = 1; i < m->length; ++i) {
 		out = fmax(out, m->elements[i]);
 	}
 
@@ -355,8 +360,8 @@ matrix *sum(matrix *m, error_return *error) {
 	out = initMatrix(newSize, newDimension, error);
 	free(newSize);
 
-	for(int i = 0; i < out->length; ++i){
-		for(int j = 0; j < m->size[m->dimension - 2]; ++j){
+	for(uint64_t i = 0; i < out->length; ++i){
+		for(uint16_t j = 0; j < m->size[m->dimension - 2]; ++j){
 			out->elements[i] += m->elements[i * m->size[m->dimension - 2] + j];
 		}
 	}
@@ -373,7 +378,7 @@ matrix *sum(matrix *m, error_return *error) {
 matrix *avg(matrix *m, error_return *error) {
 	matrix *out = sum(m, error);
 
-	for(int i = 0; i < out->length; ++i){
+	for(uint64_t i = 0; i < out->length; ++i){
 		out->elements[i] /= m->size[m->dimension - 1];
 	}
